% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity_code.r
\docType{data}
\name{BCIspeed}
\alias{BCIspeed}
\title{Animal speed data}
\format{
A dataframe with 2204 observations and 3 variables.
}
\source{
http://dx.doi.org/10.6084/m9.figshare.1160536
}
\description{
Barro Colorado Island 2008 data: speeds of animal passages past camera traps
(\code{speed}), together with species (\code{species}) and time of day (\code{time})
for each record.
}
