% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity_code.r
\docType{data}
\name{BCItime}
\alias{BCItime}
\title{Animal record time of day data}
\format{
A dataframe with 17820 observations and 2 variables.
}
\source{
http://dx.doi.org/10.6084/m9.figshare.1160536
}
\description{
Barro Colorado Island 2008 data: times of day at which animal records occured
(\code{time}), together with species (\code{species}).
}
