% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity_code.r
\name{bwcalc}
\alias{bwcalc}
\title{Calculate circular kernel bandwidth.}
\usage{
bwcalc(dat, K = 3)
}
\arguments{
\item{dat}{Numeric data vector of radian times.}

\item{K}{Integer number of values of kappa over which to maximise (see references for details).}
}
\value{
Single numeric bandwidth value.
}
\description{
Uses an optimisation procedure to calculate the circular kernel bandwidth giving the best fit to the data.
}
\details{
Mainly for internal use.
}
\references{
Ridout, M.S. & Linkie, M. (2009) Estimating overlap of daily activity patterns from camera trap data. Journal of Agricultural Biological and Environmental Statistics, 14, 322-337.
}
