% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity_code.r
\name{cmean}
\alias{cmean}
\title{Circular mean}
\usage{
cmean(x, ...)
}
\arguments{
\item{x}{A vector of radian values.}

\item{...}{Arguments passed to \code{mean}.}
}
\value{
A radian value giving mean direction.
}
\description{
Calculates the average direction of a set of radian circular values.
}
\details{
The \code{base::mean} function is use internally, and additional arguments, e.g for missing data handling, are passed to this.
}
\examples{
data(BCItime)
times <- subset(BCItime, species=="ocelot")$time*2*pi
cmean(times)
}
\seealso{
\code{\link{mean}}
}
