% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity_code.r
\name{dvonm}
\alias{dvonm}
\title{von Mises density function}
\usage{
dvonm(x, mu, k, log = FALSE)
}
\arguments{
\item{x}{numeric angles (assumed to be radian).}

\item{mu}{numeric, the mean direction of the distribution.}

\item{k}{non-negative numeric, the concentration parameter distribution (kappa).}

\item{log}{if TRUE log probabilities are returned.}
}
\value{
Probability density value(s).
}
\description{
Probability density function for the von Mises circular distribution.
}
\details{
If more than one of x, mu and k have length > 1, values are recycled.
}
\examples{
dvonm(seq(0, 2*pi, len=10), pi, 1)
}
