% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity_code.r
\docType{class}
\name{lincircmod-class}
\alias{lincircmod-class}
\title{An S4 class describing linear-circular relationships.}
\description{
An S4 class describing linear-circular relationships.
}
\section{Slots}{

\describe{
\item{\code{data}}{Object of class \code{"data.frame"}, the input data, with columns
\code{lindat} (linear data) and \code{circdat} (circular data).}

\item{\code{fit}}{Object of class \code{"data.frame"}, summary of the model fit, with columns:
\code{x}: A regular ascending sequence from 0 to 2*pi at which other columns evaluated;
\code{fit}: The linear fitted values;
\code{p}: The two tailed probability of observing the fitted values under a random (null) circular distribution;
\code{nullLCL}: The lower 95\% confidence limit of the null distribution;
\code{nullUCL}: The upper 95\% confidence limit of the null distribution.}
}}

