% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity_code_1_1.r
\docType{methods}
\name{plot,actmod-method}
\alias{plot,actmod-method}
\title{Plot activity distribution}
\usage{
\S4method{plot}{actmod}(x, xunit = c("hours", "radians"),
  yunit = c("frequency", "density"), dat = c("histogram", "rug", "none"),
  add = FALSE, dcol = 1, lcol = 2, ...)
}
\arguments{
\item{x}{Object of class \code{actmod}.}

\item{xunit}{Character string defining x-axis unit.}

\item{yunit}{Character string defining y-axis unit.}

\item{dat}{Character string defining whether to plot the data distribution and if so which style to use.}

\item{add}{Logical defining whether to create a new plot (default) or add the probability density to an existing plot (in which case no data are plotted).}

\item{dcol}{Numeric or character defining colour of data lines.}

\item{lcol}{Numeric or character defining colour of PDF lines.}

\item{...}{Additional arguments passed to internal plot call affecting only axes.}
}
\description{
Plot an activity probability distribution from a fitted \code{actmod} object.
}

