% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity_code.r
\name{plot.lincircmod}
\alias{plot.lincircmod}
\title{Plot linear-circular relationship}
\usage{
\method{plot}{lincircmod}(
  x,
  CircScale = 2 * pi,
  tlim = c(0, 1),
  fcol = "black",
  flty = 1,
  ncol = "red",
  nlty = 2,
  ...
)
}
\arguments{
\item{x}{Object of class \code{lincircmod}.}

\item{CircScale}{Single numeric value defining the plotting maximum of the circular scale.}

\item{tlim}{Numeric vector with two elements >=0 and <=1 defining the lower and upper limits at which to plot distributions; default plots the full range.}

\item{fcol, flty, ncol, nlty}{Define line colour (\code{col}) and type (\code{lty}) for fitted (\code{f}) and null (\code{n}) distributions; input types as for \code{col} and \code{lty}, see \code{\link{par}}.}

\item{...}{Additional arguments passed to the inital plot construction, affecting axes and data plot symbols.}
}
\value{
No return value, called to create a plot visualising a linear-circular relationship.
}
\description{
Plot linear against circular data along with the fitted and null confidence limit distributions from a fitted \code{lincircmod} object.
}
