% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity_code.r
\name{trigmolen}
\alias{trigmolen}
\title{title trigonometric moment length}
\usage{
trigmolen(x, p)
}
\arguments{
\item{x}{a vector of circular values, assumed to be radian.}

\item{p}{order of trigonometric moment to be computed.}
}
\value{
Trigonometric moment length of the input.
}
\description{
Calculate trigonometric moment length
}
