% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{actmod-class}
\alias{actmod-class}
\title{Activity model class.}
\description{
An S4 class describing activity models fitted to time of observation data.
}
\section{Slots}{

\describe{
\item{\code{data}}{Object of class \code{"numeric"}, the input data.}

\item{\code{wt}}{Object of class \code{"numeric"}, weights applied to the data.}

\item{\code{bw}}{Object of class \code{"numeric"}, kernel bandwidth.}

\item{\code{adj}}{Object of class \code{"numeric"}, kernel bandwidth adjustment multiplier.}

\item{\code{pdf}}{Object of class \code{"matrix"} describing fitted probability density function:
 Column 1: Sequence of radian times at which PDF evaluated (specifically seq(0,2*pi,pi/256)).
 Column 2: Corresponding circular kernel PDF values.
Additionally if errors boostrapped:
 Column 3: PDF standard error.
 Column 4: PDF lower confidence interval. Column 5: PDF upper confidence interval.}

\item{\code{act}}{Object of class \code{"numeric"} giving activity level estimate and, if errors boostrapped, standard error and 95 percent confidence limits.}
}}

