% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{plot,actmod-method}
\alias{plot,actmod-method}
\title{Plot activity distribution}
\usage{
\S4method{plot}{actmod}(x, hrs = TRUE, frq = TRUE, dat = c("histogram",
  "rug", "none"), add = FALSE, dcol = 1, lcol = 2, ...)
}
\arguments{
\item{x}{Object of class \code{actmod}.}

\item{hrs}{Logical defining whether to plot x axis as hours (default) or radians.}

\item{frq}{Logical defining whether to plot y axis as frequency (default) or probability density.}

\item{dat}{Data distribution plotting style, one of "h" (histogram, default), "r" (rug), "n" (no data plotting, PDF only).}

\item{add}{Logical defining whether to create a new plot (default) or add the probability density to an existing plot (in which case no data are plotted).}

\item{dcol}{Numeric or character defining colour of data lines.}

\item{lcol}{Numeric or character defining colour of PDF lines.}

\item{...}{Additional plotting arguments passed to internal plot call. NB axis limits are calculated internally and cannot be reset.}
}
\description{
Plot an activity probability distribution from a fitted \code{actmod} object.
}

