% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity_code_1_1.r
\docType{methods}
\name{show,actmod-method}
\alias{show,actmod-method}
\title{Show activity level estimate}
\usage{
\S4method{show}{actmod}(object)
}
\arguments{
\item{object}{Object of class \code{actmod}.}
}
\description{
Prints the \code{act} slot (activity level estimate) from an \code{actmod} object.
}

